move ProductSupplierUpdater to Update ns
